requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
feedparser==6.0.8
humanize==4.0.0
idna==3.3
python-dateutil==2.8.2
requests==2.27.1
sgmllib3k==1.0.0
six==1.16.0
urllib3==1.26.9
